# api.yaml
openapi: 3.0.2
info:
  title: DBS Programmierpraktikum - Phase 4 - API-Spezifikation
  description: >-
    In der folgenden Liste finden Sie die <b>für die Anwendergruppen
    spezifischen Grundfunktionalitäten</b>. Es müssen <b>alle folgenden
    allgemeinen Bedingungen</b> eingehalten werden:
    <ul>
      <li><b>Bei einer gültigen <code>GET</code>-Anfrage wird entweder ein JSON-Array von JSON-Objekten, falls die Ressource eine Sammlung von Entitäten darstellt, oder ein JSON-Objekt, falls die Ressource eine einzelne Entität darstellt, zurückgegeben</b> und dies mit einem Code <code>200</code> belegt. Wird die <code>GET</code>-Anfrage auf einer nicht vorhandenen Ressource ausgeführt, so wird mit Code <code>404</code> geantwortet. Ist die Ressource woanders zu finden, wird dies mit dem Code <code>301</code> und dem Link im Header belegt. Eine <code>GET</code>-Anfrage auf einer Ressource, die eine Sammlung von Entitäten darstellt, kann durch Übergabe von bestimmten Parametern und dazugehörigen Werten <b>case-insensitive gefiltert</b> werden. Dabei muss jede Entität in der Sammlung allen Parametern gleichzeitig genügen. Die Sammlung wird in der Antwort in ein JSON-Array gekapselt.</li>
...
      <li><b>Bei einer gültigen <code>DELETE</code>-Anfrage wird der Code <code>204</code> zurückgegeben.</b> Ist die Ressource nicht vorhanden, wird dies mit Code <code>404</code> belegt.</li>
...
      <li><b>Passwörter werden der Einfachheit halber im Klartext übergeben.</b></li>
...
    </ul>
...
servers:
...
paths:
...
  /kaeufe:
    get:
      tags:
        - Authentifizierter und als Kunde authorisierter Anwender
      security:
        - basic authentication: []
      summary: Gibt eigene Käufe zurück.
      description: >-
        Alle eigenen Käufe werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: name
          in: query
          description: 'Nur Käufe von einem Produkt mit **Namen** `name` werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        kaufid:
                          type: integer
                          description: '**RowID des Kaufs**.'
                        produktid:
                          type: integer
                          description: '**RowID des Produkts**.'
                        name:
                          type: string
                          description: '**Name** des Produkts.'
                        kundeid:
                          type: integer
                          description: '**RowID des Kunden**.'
                      required:
                        - kaufid
                        - produktid
                        - name
                        - kundeid
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
    post:
      tags:
        - Authentifizierter und als Kunde authorisierter Anwender
      security:
        - basic authentication: []
      summary: Kauft ein Produkt.
      description: >-
        Das Produkt mit **ProduktID** `produktid` wird mit einem **gültigen Gutscheincode** gekauft.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  properties:
                    produktid:
                      type: integer
                      description: '**RowID** des Produkts.'
                  required:
                    - produktid
                - type: object
                  properties:
                    gutscheincode:
                      type: string
                      description: '**Gutscheincode** für den Kauf.'
                  required:
                    - gutscheincode
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
  /kaeufe/{kaufid}:
    get:
      tags:
        - Authentifizierter und als Kunde authorisierter Anwender
      security:
        - basic authentication: []
      summary: Gibt einen eigenen Kauf zurück.
      description: >-
        Der Kauf mit der RowID `kaufid` wird zurückgegeben.
      parameters:
        - name: kaufid
          in: path
          description: '**RowID des Kaufs**.'
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                allOf:
                  - type: object
                    properties:
                      kaufid:
                        type: integer
                        description: '**RowID des Kaufs**.'
                    required:
                      - kaufid
                  - $ref: '#/components/schemas/kaeufe'
                  - type: object
                    properties:
                      produktid:
                        type: integer
                        description: '**RowID des Produkts**.'
                      kundeid:
                        type: integer
                        description: '**RowID** des Kunden.'
                    required:
                      - produktid
                      - kundeid
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
components:
  securitySchemes:
    basic authentication:
      type: http
      scheme: basic
  responses:
    '201':
      description: '**Created**.'
      headers:
        Location:
          description: Die URL der neuen Ressource.
          schema:
            type: string
...
    '400':
      description: '**Bad Request**.'
    '401':
      description: '**Unauthorized**.'
      headers:
        WWW-Authenticate:
          schema:
            type: string
    '403':
      description: '**Forbidden**.'
...
  schemas:
...
    kaeufe:
      type: object
      properties:
        gutscheincode:
          type: string
          description: '**Gutscheincode** des Kaufs.'
      required:
        - gutscheincode